Moved functions at top of main_gui to main_gui_functions
